+++ state/QCameraState.java
@@ -23,7 +23,7 @@ public class QCameraState extends QActorState {
 	}
 
 	@Override
-	public void onTick(int tickTime, QDirection direction) {
+	public void onTick(long tickTime, QDirection direction) {
 		// Do nothing for now
 	}
